deckgl as mapbox layer
