+++ src/app/app.module.ts
@@ -1,10 +1,10 @@
-import {BrowserModule} from "@angular/platform-browser";
-import {NgModule, OnInit} from "@angular/core";
-import {HTTP_INTERCEPTORS, HttpClientModule} from "@angular/common/http";
-import {BrowserAnimationsModule} from "@angular/platform-browser/animations";
-import {AppRoutingModule} from "./app-routing.module";
-import {AppComponent} from "./app.component";
-import {BasemapComponent} from "./basemap/basemap.component";
+import { BrowserModule } from "@angular/platform-browser";
+import { NgModule, OnInit } from "@angular/core";
+import { HTTP_INTERCEPTORS, HttpClientModule } from "@angular/common/http";
+import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
+import { AppRoutingModule } from "./app-routing.module";
+import { AppComponent } from "./app.component";
+import { BasemapComponent } from "./basemap/basemap.component";
 import {
   MatButtonModule,
   MatDialogModule,
@@ -12,19 +12,18 @@ import {
   MatSnackBarModule,
   MatTooltipModule
 } from "@angular/material";
-import {HomeComponent} from "./home/home.component";
-import {MapSettingComponent} from "./basemap/map-controls/map-setting/map-setting.component";
-import {LegendComponent} from "./basemap/map-controls/legend/legend.component";
-import {LayerControlComponent} from "./basemap/map-controls/layer-control/layer-control.component";
-import {ConfigurationService} from "./services/configuration.service";
-import {CityIOService} from "./services/cityio.service";
-import {LoginComponent} from "./login/login.component";
-import {fakeBackendProvider} from "./interceptors/fake-backend";
-import {JwtInterceptor} from "./interceptors/jwt.interceptor";
-import {ReactiveFormsModule} from "@angular/forms";
-import {ErrorInterceptor} from "./interceptors/error.interceptor";
-import {ExitEditorDialog} from "./dialogues/exit-editor-dialog";
-import { DeckglComponent } from './basemap/deckgl/deckgl.component';
+import { HomeComponent } from "./home/home.component";
+import { MapSettingComponent } from "./basemap/map-controls/map-setting/map-setting.component";
+import { LegendComponent } from "./basemap/map-controls/legend/legend.component";
+import { LayerControlComponent } from "./basemap/map-controls/layer-control/layer-control.component";
+import { ConfigurationService } from "./services/configuration.service";
+import { CityIOService } from "./services/cityio.service";
+import { LoginComponent } from "./login/login.component";
+import { fakeBackendProvider } from "./interceptors/fake-backend";
+import { JwtInterceptor } from "./interceptors/jwt.interceptor";
+import { ReactiveFormsModule } from "@angular/forms";
+import { ErrorInterceptor } from "./interceptors/error.interceptor";
+import { ExitEditorDialog } from "./dialogues/exit-editor-dialog";
 
 @NgModule({
   declarations: [
@@ -35,8 +34,7 @@ import { DeckglComponent } from './basemap/deckgl/deckgl.component';
     LegendComponent,
     LayerControlComponent,
     LoginComponent,
-    ExitEditorDialog,
-    DeckglComponent
+    ExitEditorDialog
   ],
   imports: [
     BrowserModule,
@@ -55,8 +53,8 @@ import { DeckglComponent } from './basemap/deckgl/deckgl.component';
     CityIOService,
     ConfigurationService,
 
-    {provide: HTTP_INTERCEPTORS, useClass: JwtInterceptor, multi: true},
-    {provide: HTTP_INTERCEPTORS, useClass: ErrorInterceptor, multi: true},
+    { provide: HTTP_INTERCEPTORS, useClass: JwtInterceptor, multi: true },
+    { provide: HTTP_INTERCEPTORS, useClass: ErrorInterceptor, multi: true },
 
     // provider used to create fake backend
     fakeBackendProvider
@@ -65,6 +63,5 @@ import { DeckglComponent } from './basemap/deckgl/deckgl.component';
   bootstrap: [AppComponent]
 })
 export class AppModule implements OnInit {
-  ngOnInit() {
-  }
+  ngOnInit() {}
 }
