+++ src/app/basemap/basemap.component.html
@@ -1,9 +1,7 @@
 <div id="map-container" class="map-container">
   <div class="basemap" id="basemap"></div>
-  <app-deckgl></app-deckgl>
 
   <div *ngIf="isShowMenu" class="modal-container container-left-bottom">
-    <!--<app-legend [layer]="mapKeyLayer" [visible]="mapKeyVisible" (close)="mapKeyVisible = false"></app-legend>-->
     <app-layer-control
       [layers]="layers"
       (toggleLayer)="toggleLayer()"
