--- src/app/basemap/deckgl/deckgl.component.html
@@ -1,1 +1,0 @@
-<canvas [id]="glId"></canvas>
