--- src/app/basemap/deckgl/deckgl.component.scss
@@ -1,3 +1,0 @@
-canvas {
-  background-color: rgba(0, 0, 0, 0);
-}
