--- src/app/basemap/deckgl/deckgl.component.ts
@@ -1,100 +1,0 @@
-import { Component } from "@angular/core";
-import {
-  AfterViewInit,
-  Input,
-  Output,
-  NgZone,
-  EventEmitter
-} from "@angular/core";
-
-import { Deck, MapController } from "@deck.gl/core";
-import { TripsLayer } from "@deck.gl/geo-layers";
-
-const DATA_URL = {
-  TRIPS:
-    "https://raw.githubusercontent.com/uber-common/deck.gl-data/master/examples/trips/trips.json" // eslint-disable-line
-};
-
-@Component({
-  selector: "app-deckgl",
-  templateUrl: "./deckgl.component.html",
-  styleUrls: ["./deckgl.component.scss"]
-})
-export class DeckglComponent implements AfterViewInit {
-  @Input() glId: string;
-  @Input() width: number;
-  @Input() height: number;
-  @Output() viewPortChange = new EventEmitter();
-  private deckgl;
-
-  private controller;
-
-  private viewport = {
-    latitude: 40.75,
-    longitude: -74,
-    zoom: 12,
-    pitch: 45,
-    bearing: 0
-  };
-
-  constructor(public ngZone: NgZone) {}
-
-  ngAfterViewInit() {
-    this.ngZone.runOutsideAngular(() => {
-      this.deckgl = new Deck({
-        ...this.viewport,
-        layers: [],
-        canvas: document.getElementById(this.glId),
-        initialViewState: this.viewport
-        // controller: true
-      });
-
-      this.controller = new MapController({
-        canvas: this.deckgl.canvas,
-        ...this.viewport,
-        onViewStateChange: this.onViewportChange,
-        width: this.width || 500,
-        height: this.height || 500
-      });
-
-      setInterval(() => {
-        this.renderLayers();
-      });
-    });
-  }
-
-  onViewportChange = (viewport): void => {
-    this.viewport = viewport;
-    this.deckgl.setProps(viewport);
-    this.controller.setProps(viewport);
-    this.renderLayers();
-    this.ngZone.runOutsideAngular(() => {
-      this.viewPortChange.emit(viewport);
-    });
-  };
-
-  renderLayers() {
-    let loopLength = 1800;
-    let animationSpeed = 50;
-    const timestamp = Date.now() / 1000;
-    const loopTime = loopLength / animationSpeed;
-    let time = ((timestamp % loopTime) / loopTime) * loopLength;
-    let trailLength = 300;
-
-    const tripLayer = new TripsLayer({
-      id: "trips",
-      data: DATA_URL.TRIPS,
-      getPath: d => d.segments,
-      getColor: d => (d.vendor === 0 ? [255, 50, 255] : [60, 150, 255]),
-      opacity: 0.5,
-      widthMinPixels: 2,
-      rounded: true,
-      trailLength,
-      currentTime: time
-    });
-    // then put this updated layer into deck
-    this.deckgl.setProps({
-      layers: [tripLayer]
-    });
-  }
-}
