+++ src/app/config.json
@@ -1,6 +1,6 @@
 {
   "mapCenter": [10.0143909533867, 53.53128461384861],
-  "mapZoom": 16,
+  "mapZoom": 10,
   "bearing": 34.8,
   "pitch": 30,
   "gridZoom": 16.3,
